manifests/middleware/requirements.txt
# Core Flask web framework and server
flask==2.3.3
gunicorn==21.2.0

# HTTP requests and API communication
requests==2.31.0

# JWT token handling for authentication
PyJWT==2.8.0

# YAML parsing for configuration
pyyaml==6.0.1

# Feature flag provider SDKs
# Install with: pip install -r requirements-[provider].txt
# Or use setup script which will configure automatically

# WebSocket support for real-time features
flask-socketio==5.3.6
python-socketio==5.8.0

# CORS support for cross-origin requests
flask-cors==4.0.0

# Password hashing
bcrypt==4.0.1

# OpenTelemetry instrumentation and exporters
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-socketio==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation-logging==0.42b0